Related: bug #632425
fix password & CryptoManger initialization

These are a couple of simple issues which I discovered during the tomcat
porting. Since they are mostly independent of the tomcat changes I'd like to
keep them separate and get them in first.

The changes are:

1) Make CMSEngine.getPasswordStore() a synchronized method.

During testing I had discovered two threads were both calling this 
method at essentially the same time trashing the state which lead to 
password failures.

2) Remove redundant duplicated code for initializing the password store, 
now CMSEngine.getPasswordStore() is the sole owner of the logic to 
perform this action.

3) Initialize CryptoManager before first use. We had been relying on a 
side effect in tomcat 5 for CryptoManager initialization where tomcatjss 
was doing the CryptoManager initialization prior to our first use of the 
CryptoManager. Tomcat 6 has modified when the connection objects first 
get created (which was what was kicking off the CryptoManager 
initialization). The patch adds the same code for initializing the 
CryptoManger as is in tomcatjss. We now check for CryptoManager 
initialization *prior* to our first use of it and if and only if it 
hasn't been initialized yet we do so.