Add run_command() utility
Many of the existing functions were using the backtic operator to run
a shell command and then tested if the length returned string was zero
to determine success. This is incorrect for two reasons:

1) the backtick operator discards stderr and returns only stdout, thus
if the command did generate a message because of an error it would
likely do so on stderr not stdout and thus the test for output is
performed on the wrong stream.

2) the presence or absense of output is not the proper way to
determine if a shell command succeeded, one needs to test the exit
status of the command.

This patch adds the run_command() function which will be utilized in
subsequent patches.