Parameter list style fix
The preferred canonical style for Perl subroutine parameter lists is
to write the parameter list as an actual parameter list prefixed by
the "my" lexical modifier and initialized from the @_ parameter
array. The parameter list should be the first line in the sub. This
makes it much easier to read the code because a parameter list looks
like a parameter list. This improves readability and follows widely
adopted style conventions. For example:

sub foo
{
    my ($able, $baker) = @_;
}

Thus the signature for this subroutine is:

foo($able, $baker)

The patch also removes comments of the form

# arg0 is able
# arg1 is baker

Which are silly when you write the parameter list to look like a
parameter list because it's self evident what the parameters
are. Comments like that clutter the code, decreases readability and
does not add any information content.