Fix ampersand function calls
Some functions were being called with the deprecated ampersand
syntax. In Perl the & prefix operator indicates the expression is to
be interpreted as a function, e.g. &foo means foo is a function and if
foo was followed by a list then it means call the function foo. The
list can be parenthesized or not, it could just be comma separated
expressions. Calling functions with this syntax is a hold over from
earlier versions of Perl, but modern Perl has much cleaner syntax
where function calls look like they do in other languages, an
identifier followed by parenthesis. This is the calling style used in
most of the rest of the code. This patch just unifies the calling
syntax so it's consistent and more readable. Also the patch cleans up
the function definition, some of the functions had been defined with
an empty formal parameter list, but that conflicts with function
prototyping introduced in modern Perl, an empty formal parameter list
states the function takes no arguments. It only worked previously
because when the (deprecated) ampersand operator was applied to the
identifier it defeated prototype checking.